+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Update to rails 6.0.0.beta1
   - Update to carrierwave 1.3.1
   - Update to bootstrap 4.2.1
+  - Update to font-awesome-sass 5.6.1
 
 ## 4.0.0 (November 7, 2018)